+++ tools/tests/functional/unit_tests/algorithmic/core_pack_be_idx.cpp
@@ -103,8 +103,8 @@ QPL_UNIT_API_ALGORITHMIC_TEST(qplc_pack_index_be_8u16u, base) {
             uint32_t       ref_index     = 0U;
             const uint32_t status        = (uint32_t)qplc_pack_vector(fun_indx_pack_be_index_8u16u)(
                     source.data(), length, &p_dst_16u, length_dst, &index);
-            const uint32_t ref_status = (uint32_t)ref_qplc_pack_be_index_8u16u(source.data(), length, &p_ref_dst_16u,
-                                                                               length_dst, &ref_index);
+            const uint32_t ref_status =
+                    ref_qplc_pack_be_index_8u16u(source.data(), length, &p_ref_dst_16u, length_dst, &ref_index);
             ASSERT_EQ(status, ref_status);
             ASSERT_EQ(index, ref_index);
             ASSERT_EQ((p_dst_16u - (uint8_t*)destination.data()), (p_ref_dst_16u - (uint8_t*)reference.data()));
@@ -122,8 +122,8 @@ QPL_UNIT_API_ALGORITHMIC_TEST(qplc_pack_index_be_8u16u, base) {
             uint32_t       ref_index     = UINT16_MAX - 32U;
             const uint32_t status        = (uint32_t)qplc_pack_vector(fun_indx_pack_be_index_8u16u)(
                     source.data(), length, &p_dst_16u, length_dst, &index);
-            const uint32_t ref_status = (uint32_t)ref_qplc_pack_be_index_8u16u(source.data(), length, &p_ref_dst_16u,
-                                                                               length_dst, &ref_index);
+            const uint32_t ref_status =
+                    ref_qplc_pack_be_index_8u16u(source.data(), length, &p_ref_dst_16u, length_dst, &ref_index);
             ASSERT_EQ(status, ref_status);
             ASSERT_EQ(index, ref_index);
             ASSERT_EQ((p_dst_16u - (uint8_t*)destination.data()), (p_ref_dst_16u - (uint8_t*)reference.data()));
@@ -156,8 +156,8 @@ QPL_UNIT_API_ALGORITHMIC_TEST(qplc_pack_index_be_8u32u, base) {
             uint32_t       ref_index     = 0U;
             const uint32_t status        = (uint32_t)qplc_pack_vector(fun_indx_pack_be_index_8u32u)(
                     source.data(), length, &p_dst_32u, length_dst, &index);
-            const uint32_t ref_status = (uint32_t)ref_qplc_pack_be_index_8u32u(source.data(), length, &p_ref_dst_32u,
-                                                                               length_dst, &ref_index);
+            const uint32_t ref_status =
+                    ref_qplc_pack_be_index_8u32u(source.data(), length, &p_ref_dst_32u, length_dst, &ref_index);
             ASSERT_EQ(status, ref_status);
             ASSERT_EQ(index, ref_index);
             ASSERT_EQ((p_dst_32u - (uint8_t*)destination.data()), (p_ref_dst_32u - (uint8_t*)reference.data()));
@@ -175,8 +175,8 @@ QPL_UNIT_API_ALGORITHMIC_TEST(qplc_pack_index_be_8u32u, base) {
             uint32_t       ref_index     = UINT32_MAX - 32U;
             const uint32_t status        = (uint32_t)qplc_pack_vector(fun_indx_pack_be_index_8u32u)(
                     source.data(), length, &p_dst_32u, length_dst, &index);
-            const uint32_t ref_status = (uint32_t)ref_qplc_pack_be_index_8u32u(source.data(), length, &p_ref_dst_32u,
-                                                                               length_dst, &ref_index);
+            const uint32_t ref_status =
+                    ref_qplc_pack_be_index_8u32u(source.data(), length, &p_ref_dst_32u, length_dst, &ref_index);
             ASSERT_EQ(status, ref_status);
             ASSERT_EQ(index, ref_index);
             ASSERT_EQ((p_dst_32u - (uint8_t*)destination.data()), (p_ref_dst_32u - (uint8_t*)reference.data()));
